+++ app/Lib/Migration/Migration.php
@@ -5,6 +5,7 @@
  * @copyright (c) 2015-present Bolt Softwares Pvt Ltd
  * @licence GNU Affero General Public License http://www.gnu.org/licenses/agpl-3.0.en.html
  */
+App::uses('HttpSocket', 'Network/Http');
 App::uses('MigrationVersion', 'Migrations.Lib');
 
 class Migration extends CakeObject {